refactor: 회원 여부에 따른 분기 처리 상세화 - fail : 서버 통신 에러 - not found : 회원 아님 - success : 회원임
issue:#12
