+++ app/src/main/java/com/sw/wordgarden/app/AuthCodeHandlerActivity.kt
@@ -1,4 +1,4 @@
-package com.sw.wordgarden.Oauth
+package com.sw.wordgarden.app
 
 import android.content.ContentValues.TAG
 import android.os.Bundle
