+++ app/src/main/java/com/sw/wordgarden/presentation/model/UserCheckEvent.kt
@@ -1,0 +1,9 @@
+package com.sw.wordgarden.presentation.model
+
+import androidx.annotation.StringRes
+
+sealed interface UserCheckEvent {
+    data object Success: UserCheckEvent
+    data object NotFound: UserCheckEvent
+    data class Failure(@StringRes val msg: Int): UserCheckEvent
+}
