+++ app/src/main/res/values/strings.xml
@@ -7,6 +7,8 @@
     <string name="login_kakao">카카오로 계속하기</string>
     <string name="login_start">가장 편한 방법으로 시작하세요!</string>
     <string name="login_msg_fail_check">회원 정보 조회에 실패했습니다.</string>
+    <string name="login_msg_result_null">회원 정보가 없습니다.</string>
+    <string name="login_msg_not_user">회원 가입이 필요합니다.</string>
     <string name="onboarding_input_nickname">닉네임을 입력하세요</string>
     <string name="onboarding_next">다음</string>
     <string name="onboarding_msg_fill">닉네임을 입력해주세요.</string>
